Improved port handling

+++ inc/upnp/upnp.client.h
@@ -47,11 +47,11 @@ public:
     Client2(uv::Loop& loop);
     virtual ~Client2();
 
-    virtual void initialize(const std::string& interfaceName, int32_t port);
+    virtual void initialize(const std::string& interfaceName, uint16_t port);
     virtual void uninitialize();
 
     virtual std::string getIpAddress() const;
-    virtual int32_t getPort() const;
+    virtual uint16_t getPort() const;
 
     virtual void subscribeToService(const std::string& publisherUrl, std::chrono::seconds timeout, std::function<void(int32_t status, std::string subId, std::chrono::seconds timeout)> cb);
     virtual void unsubscribeFromService(const std::string& publisherUrl, const std::string& subscriptionId, std::function<void(int32_t status)> cb);
